Auth / posts / comments / users BACK + FRONT

+++ BACKEND/controllers/admin.js
@@ -1,3 +1,4 @@
+const selfOrAdmin = require('../classes/selfOrAdmin');
 const models = require('../models');
 
 
@@ -10,13 +11,20 @@ exports.create = (request, response, next) => {
         return response.status(400).json({message: "Aucun ID n'a été envoyé"})
     }
 
-    models.Admin.create({ UserId: user_id })
-        .then(() => {
-            response.status(200).json({message: "Admin rajouté avec succès"});
+    selfOrAdmin(request)
+        .then(check => {
+            if(check == "admin") {
+                models.Admin.create({ UserId: user_id })
+                    .then(() => {
+                        return response.status(200).json({message: "Admin rajouté avec succès"}); 
+                    })
+                    .catch(error => response.status(500).json({message: error.message}))
+            }
+            else {
+                return response.status(403).json({message: "Accès refusé"});
+            }
         })
-        .catch(error => {
-            response.status(500).json({message: error});
-        })
+        .catch(error => response.status(500).json({message: error.message}))
 }
 
 exports.delete = (request, response, next) => {
@@ -27,12 +35,37 @@ exports.delete = (request, response, next) => {
     if(user_id == null || isNaN(user_id) ) {
         return response.status(400).json({message: "Aucun ID n'a été envoyé"})
     }
-    
-    models.Admin.destroy({ where: { UserId: user_id }})
-        .then(() => {
-            response.status(200).json({message: "Admin supprimé avec succès"});
+
+    selfOrAdmin(request)
+        .then(check => {
+            if(check == "admin") {
+                models.Admin.destroy({ where: { UserId: user_id }})
+                    .then(() => {
+                        response.status(200).json({message: "Admin supprimé avec succès"});
+                    })
+                    .catch(error => response.status(500).json({message: error.message}))
+            }
+            else {
+                return response.status(403).json({message: "Accès refusé"});
+            }
         })
-        .catch(error => {
-            response.status(500).json({message: error.message});
+        .catch(error => response.status(500).json({message: error.message}))
+}
+
+exports.isAdmin = (request, response, next) => {
+    // Request : null
+    // Response : {isAdmin: boolean}
+    let user_id = request.params.id;
+    let isAdmin = false;
+
+    if(user_id == null || isNaN(user_id) ) {
+        return response.status(400).json({message: "Aucun ID n'a été envoyé"})
+    }
+
+    models.Admin.findOne({where: {UserId: user_id}})
+        .then(admin => {
+            if(admin) {isAdmin = true;}
+            response.status(200).json({isAdmin: isAdmin});
         })
+        .catch(error => response.status(500).json({message: error.message}));
 }
