+++ BACKEND/controllers/comment.js
@@ -1,23 +1,24 @@
+const selfOrAdmin = require('../classes/selfOrAdmin');
 const models  = require('../models');
 
 exports.create = (request, response, next) => {
-    // Request : {content: String, topic_id: Integer, user_id: Integer}
+    // Request : {content: String, post_id: Integer, user_id: Integer}
     // Response : {message: String}
     
     let content = request.body.content;
-    let topic_id = request.body.topic_id;
+    let post_id = request.body.post_id;
     let user_id = request.body.user_id;
 
-    if(content == null || topic_id == null || user_id == null) {
+    if(content == null || post_id == null || user_id == null) {
         return response.status(400).json({ error: 'Il manque un parametre'});
     }
 
     models.Comment.create({
         content: content,
-        TopicId: topic_id,
+        PostId: post_id,
         UserId: user_id
     })
-        .then(() => response.status(200).json({message: "Commentaire créé avec succès !"}))
+        .then(comment => response.status(200).json({message: "Commentaire créé avec succès !"}))
         .catch(error => response.status(500).json({error: error.message}))
 }
 
@@ -47,7 +48,7 @@ exports.getOne = (request, response, next) => {
 
 exports.update = (request, response, next) => {
     // Request : {content: String?}
-    // Response : {comment: Comment}
+    // Response : {message: String}
 
     let comment_id = request.params.id;
 
@@ -55,8 +56,14 @@ exports.update = (request, response, next) => {
         return response.status(400).json({message: "Aucun ID n'a été envoyé"})
     }
 
-    models.Comment.update({ ...request.body },{ where: {id: comment_id }})
-        .then(comment => response.status(200).json({comment: comment}))
+    selfOrAdmin(request, "Comment")
+        .then(check => {
+            if(check == "admin" || check == "self") {
+                models.Comment.update({ ...request.body },{ where: {id: comment_id }})
+                    .then(() => response.status(200).json({message: "Commentaire modifié avec succès"}))
+                    .catch(error => response.status(500).json({error: error.message}));
+            }
+        })
         .catch(error => response.status(500).json({error: error.message}));
 }
 
@@ -70,7 +77,13 @@ exports.delete = (request, response, next) => {
         return response.status(400).json({message: "Aucun ID n'a été envoyé"})
     }
 
-    models.Comment.destroy({where: { id: comment_id}})
-        .then(() => response.status(200).json({message: "Commentaire supprimé avec succès"}))
-        .catch(error => response.status(500).json({error: error.message}))
+    selfOrAdmin(request, "Comment")
+        .then(check => {
+            if(check == "admin" || check == "self") {
+                models.Comment.destroy({where: { id: comment_id}})
+                    .then(() => response.status(200).json({message: "Commentaire supprimé avec succès"}))
+                    .catch(error => response.status(500).json({error: error.message}))
+            }
+        })
+        .catch(error => response.status(500).json({error: error.message}));
 }
