+++ BACKEND/controllers/user.js
@@ -1,3 +1,4 @@
+const selfOrAdmin = require('../classes/selfOrAdmin');
 const models  = require('../models');
 
 exports.getAll = (request, response, next) => {
@@ -11,7 +12,7 @@ exports.getAll = (request, response, next) => {
 
 exports.getOne = (request, response, next) => {
     // Request : null
-    // Response : {user: User}
+    // Response : {user: User[]}
     
     let user_id = request.params.id;
 
@@ -20,7 +21,7 @@ exports.getOne = (request, response, next) => {
     }
 
     models.User.findOne({where: { id: user_id}})
-        .then(user => response.status(200).json({user: user}))
+        .then(user => response.status(200).json({user: [user]}))
         .catch(error => response.status(500).json({error: error.message}))
 }
 
@@ -28,14 +29,20 @@ exports.update = (request, response, next) => {
     // Request : {name: String?, first_name: String?, Email: String?, password: String?}
     // Response : {user: User}
 
-    let user_id = request.params.id;
+    let user_id = request.params.userId;
 
     if(user_id == null || isNaN(user_id) ) {
         return response.status(400).json({message: "Aucun ID n'a été envoyé"})
     }
 
-    models.User.update({ ...request.body },{ where: {id: user_id }})
-        .then(user => response.status(200).json({user: user}))
+    selfOrAdmin(request)
+        .then(check => {
+            if(check == "admin" || check == "self") {
+                models.User.update({ ...request.body },{ where: {id: user_id }})
+                    .then(user => response.status(200).json({user: user}))
+                    .catch(error => response.status(500).json({error: error.message}));
+            }
+        })
         .catch(error => response.status(500).json({error: error.message}));
 }
 
@@ -43,13 +50,21 @@ exports.delete = (request, response, next) => {
     // Request : null
     // Response : {Message: String}
     
-    let user_id = request.params.id;
+    let user_id = request.params.userId;
 
     if(user_id == null || isNaN(user_id) ) {
         return response.status(400).json({message: "Aucun ID n'a été envoyé"})
     }
-
-    models.User.destroy({where: { id: user_id}})
-        .then(() => response.status(200).json({message: "Utilisateur supprimé avec succès"}))
-        .catch(error => response.status(500).json({error: error.message}))
+    selfOrAdmin(request)
+        .then(check => {
+            if(check == "admin" || check == "self") {
+                models.User.destroy({where: {id: user_id}})
+                    .then(() => response.status(200).json({message: "Utilisateur supprimé avec succès"}))
+                    .catch(error => response.status(500).json({error: error.message}))
+            }
+            else {
+                return response.status(403).json({error: "Accès refusé."})
+            }
+        })
+        .catch(error => response.status(500).json({error: error.message}));
 }
